check pointer for nil in association release
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@7880 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ Binaries/dcmtk-source/dcmnet/assoc.cc
@@ -358,7 +358,9 @@ ASC_destroyAssociationParameters(T_ASC_Parameters ** params)
   * association terminates. 
   */
 {
-
+	if( params == nil) return EC_Normal;
+	if( *params == nil) return EC_Normal;
+	
     /* free the elements in the requested presentation context list */
     destroyPresentationContextList(
         &((*params)->DULparams.requestedPresentationContext));
